Extract Spot out of Sudoku.
