+++ HW3-SudokuSolver/src/main/java/assign3/Sudoku.java
@@ -7,132 +7,7 @@ import java.util.*;
  * This project is based on the assignment given in CS108 Stanford.
  */
 public class Sudoku {
-	
-	/* "Spot" inner class that represents a single spot
-	 * on the grid of the Sudoku game.
-	 * Each Spot object knows its place on the Sudoku grid as
-	 * it stores the row and column number as fields.
-	 */
-	private class Spot implements Comparable<Spot> {
-		
-		/* Properties/fields of each individual Spot */
-		private final int row, col;
-		private int value;
-		private int part;
-		
-		/* Stores all possible values for a Spot that is empty
-		 * according to the rules of the game */
-		private Set<Integer> possibleValues;
-		
-		Spot(int row, int col, int value) {
-			this.row = row;
-			this.col = col;
-			this.value = value;
-			this.part = getPart(row, col);
-		
-			possibleValues = new HashSet<>();
-		}
-		
-		Spot(Spot s) {
-			this(s.row, s.col, s.value);
-			this.part = s.part;
-			possibleValues = new HashSet<>(s.possibleValues);
-		}
-		
-		/* Sets the value for this Spot on the Solution Grid (solutionGrid) */
-		void setValue(int val) {
-			value = val;
-		}
-		
-		void setEmpty() {
-			value = 0;
-		}
-		
-		/* Returns the value of this Spot */
-		int getValue() {
-			return value;
-		}
-		
-		/* Returns the part of the grid where this Spot belongs */
-		int getPartForSpot() {
-			return part;
-		}
-		
-		/* Returns true iff this Spot is not filled */
-		boolean isEmpty() {
-			return value == 0;
-		}
-		
-		/* Returns a HashSet of all legal values that can be 
-		 * filled in this Spot.
-		 */
-		Set<Integer> getPossibleValues() {
-			if (value != 0) return null;
 
-			/* temporarily assign all 9 numbers */
-			for (int i = 1; i <= Sudoku.SIZE; i++)
-				possibleValues.add(i);
-			
-			/* Remove all the values that cannot be placed at this Spot */
-			possibleValues.removeAll(valInRows.get(row));
-			possibleValues.removeAll(valInCols.get(col));
-			possibleValues.removeAll(valInParts.get(part));
-			return possibleValues;
-		}
-		
-		/* Updates the Spot on the solution grid with the row and col of 
-		 * this Spot with the current value that this spot holds.
-		 */
-		void updateValueInGrid() {
-			solutionGrid[row][col].value = value;
-		}
-		
-		/* Returns the row number this Spot belongs to */
-		int getRow() {
-			return row;
-		}
-
-		/* Returns the column number this Spot belongs to */
-		int getCol() {
-			return col;
-		}
-		
-		@Override
-		public int compareTo(Spot that) {
-			return this.possibleValues.size() - that.possibleValues.size();
-		}
-		
-		/* Two Spots are equal if their row and column number are equal */
-		@Override
-		public boolean equals(Object o) {
-			if (o == null) return false;
-			if (!(o instanceof Spot)) return false;
-			
-			Spot that = (Spot) o;
-			return this.row == that.row && this.col == that.col;
-		}
-		
-		@Override
-		public int hashCode() {
-			return possibleValues.size() * 25;
-		}
-		
-		@Override
-		public String toString() {
-			return String.valueOf(value);
-//			return possibleValues.toString();
-		}
-		
-		/* Helper method that returns the Part in which the 
-		 * coordinates x and y belong on the grid. 
-		 */
-		private int getPart(int x, int y) {
-		    int xGroup = x / 3;
-		    int yGroup = y / 3;
-		    return xGroup * 3 + yGroup;
-		}
-	} // End of Spot class
-	
 	/* Member variables for the puzzle and solution grids */
 	private final Spot[][] puzzleGrid;
 	private final Spot[][] solutionGrid;
@@ -204,8 +79,13 @@ public class Sudoku {
 	"6 0 0 0 2 1 0 4 0",
 	"0 0 0 5 3 0 9 0 0",
 	"0 3 0 0 0 0 0 5 1");
-	
 
+	/* Updates the Spot on the solution grid with the row and col of
+     * this Spot with the current value that this spot holds.
+     */
+	void updateValueInGrid(Spot spot) {
+		solutionGrid[spot.getRow()][spot.getCol()] = spot;
+	}
 	
 	public static final int SIZE = 9;  // size of the whole 9x9 puzzle
 	public static final int PART = 3;  // size of each 3x3 part
@@ -386,23 +266,25 @@ public class Sudoku {
 		Spot currentSpot = new Spot(emptySpots.get(index));
 		
 		/* Try all the possible values for this empty Spot */
-		for (int value : currentSpot.possibleValues) {
+		if(currentSpot.getPossibleValues(valInRows, valInCols, valInParts).isPresent()) {
+			for (int value : currentSpot.getPossibleValues(valInRows, valInCols, valInParts).get()) {
 			
 			/* Check if the value is valid according to the current 
 			 * state of the grid */
-			if (valueIsValid(value, currentSpot)) {
-				currentSpot.setValue(value);
-				updateGridStateWithValue(value, currentSpot);
-				
-				solvedSpots.add(new Spot(currentSpot));
-				
-				int newIndex = index + 1;
-				solveSudoku(emptySpots, solvedSpots, newIndex);
+				if (valueIsValid(value, currentSpot)) {
+					currentSpot.setValue(value);
+					updateGridStateWithValue(value, currentSpot);
+
+					solvedSpots.add(new Spot(currentSpot));
+
+					int newIndex = index + 1;
+					solveSudoku(emptySpots, solvedSpots, newIndex);
 				
 				/* Backtrack when the method above returns */
-				emptySpots.get(index).setEmpty();
-				solvedSpots.remove(currentSpot);
-				updateGridStateWithValue(value, currentSpot);
+					emptySpots.get(index).setEmpty();
+					solvedSpots.remove(currentSpot);
+					updateGridStateWithValue(value, currentSpot);
+				}
 			}
 		}
 	}
@@ -420,7 +302,7 @@ public class Sudoku {
 		}
 		
 		for (Spot spot : solvedSpots)
-			spot.updateValueInGrid();
+			updateValueInGrid(spot);
 	}
 
 	/* Checks if the given value is valid for the given currentSpot by 
@@ -473,7 +355,7 @@ public class Sudoku {
 			for (int j = 0; j < SIZE; j++) {
 				Spot thisSpot = puzzleGrid[i][j];
 				if (thisSpot.isEmpty()) {
-					thisSpot.getPossibleValues();
+					thisSpot.getPossibleValues(valInRows, valInCols, valInParts);
 					result.add(thisSpot);
 				}
 			}
